gate/creatures/java_adapter.py
#!/usr/bin/env python3
"""
Java Adapter: Interoperability Layer for Java Language

This module provides an adapter for integrating Java language
capabilities within the system.
"""

from typing import Any, Dict, Callable
from dataclasses import dataclass, field
import jpype
import jpype.imports

@dataclass
class JavaComponent:
    """
    Represents a Java language component.
    """
    name: str
    version: str = "17"
    capabilities: Dict[str, Any] = field(default_factory=dict)
    classpath: str = None
    
    def initialize_jvm(self) -> bool:
        """
        Initialize the Java Virtual Machine (JVM).
        
        Returns:
            bool: True if JVM initialization was successful, False otherwise.
        """
        try:
            if not jpype.isJVMStarted():
                jpype.startJVM(classpath=self.classpath)
            return True
        except Exception as e:
            print(f"JVM Initialization Error: {e}")
            return False
    
    def load_java_class(self, class_name: str) -> Any:
        """
        Load a Java class dynamically.
        
        Args:
            class_name (str): Fully qualified name of the Java class.
        
        Returns:
            Any: Loaded Java class or None if loading fails.
        """
        if not self.initialize_jvm():
            return None
        
        try:
            return jpype.JClass(class_name)
        except Exception as e:
            print(f"Java Class Loading Error: {e}")
            return None
    
    def execute_java_method(self, class_name: str, method_name: str, *args, **kwargs) -> Any:
        """
        Execute a method of a Java class.
        
        Args:
            class_name (str): Fully qualified name of the Java class.
            method_name (str): Name of the method to execute.
            *args: Positional arguments for the method.
            **kwargs: Keyword arguments for the method.
        
        Returns:
            Any: Result of the method execution.
        """
        java_class = self.load_java_class(class_name)
        if not java_class:
            return None
        
        try:
            instance = java_class()
            method = getattr(instance, method_name)
            return method(*args, **kwargs)
        except Exception as e:
            print(f"Java Method Execution Error: {e}")
            return None

class JavaAdapter:
    """
    Adapter for managing Java language components and interactions.
    """
    def __init__(self):
        """
        Initialize the JavaAdapter.
        """
        self.components: Dict[str, JavaComponent] = {}
    
    def register_component(self, component: JavaComponent) -> None:
        """
        Register a Java component.
        
        Args:
            component (JavaComponent): Java component to register.
        """
        self.components[component.name] = component
    
    def get_component(self, name: str) -> JavaComponent:
        """
        Retrieve a registered Java component.
        
        Args:
            name (str): Name of the component.
        
        Returns:
            JavaComponent: Retrieved Java component.
        
        Raises:
            KeyError: If the component is not found.
        """
        return self.components[name]
    
    def create_enterprise_component(self, name: str, classpath: str = None, capabilities: Dict[str, Any] = None) -> JavaComponent:
        """
        Create an enterprise Java component.
        
        Args:
            name (str): Name of the component.
            classpath (str, optional): Classpath for the Java component.
            capabilities (Dict[str, Any], optional): Component capabilities.
        
        Returns:
            JavaComponent: Created Java component.
        """
        component = JavaComponent(
            name=name,
            version="17",
            classpath=classpath,
            capabilities=capabilities or {}
        )
        self.register_component(component)
        return component

def create_default_java_adapter() -> JavaAdapter:
    """
    Create a default JavaAdapter with some example components.
    
    Returns:
        JavaAdapter: Configured Java adapter.
    """
    adapter = JavaAdapter()
    
    # Example enterprise components
    enterprise_component = adapter.create_enterprise_component(
        "EnterpriseFramework",
        capabilities={
            "dependency_injection": True,
            "orm_support": True,
            "microservices_architecture": True
        }
    )
    
    web_component = adapter.create_enterprise_component(
        "WebServices",
        capabilities={
            "rest_api": True,
            "spring_boot_integration": True,
            "security_framework": True
        }
    )
    
    return adapter 